# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sistema de Consultas Médicas",
        "description": "Documentação da API do Sistema de Consultas Médicas.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/medicos": {
            "get": {
                "tags": [
                    "Médicos"
                ],
                "summary": "Lista todos os médicos",
                "description": "Retorna todos os médicos cadastrados.",
                "operationId": "6f984903f8598bc5072db084bbb83535",
                "responses": {
                    "200": {
                        "description": "Lista de médicos retornada com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "data": [
                                        {
                                            "id": 1,
                                            "nome": "Dra. Ana Martins",
                                            "crm": "CRM001",
                                            "created_at": "2026-07-22T23:07:46.000000Z",
                                            "updated_at": "2026-07-22T23:07:46.000000Z"
                                        },
                                        {
                                            "id": 2,
                                            "nome": "Dr. Carlos Souza",
                                            "crm": "CRM002",
                                            "created_at": "2026-07-22T23:07:46.000000Z",
                                            "updated_at": "2026-07-22T23:07:46.000000Z"
                                        },
                                        {
                                            "id": 3,
                                            "nome": "Dr. Felipe Ramos",
                                            "crm": "CRM003",
                                            "created_at": "2026-07-22T23:07:46.000000Z",
                                            "updated_at": "2026-07-22T23:07:46.000000Z"
                                        },
                                        {
                                            "id": 4,
                                            "nome": "Dra. Fernanda Lopes",
                                            "crm": "CRM004",
                                            "created_at": "2026-07-22T23:07:46.000000Z",
                                            "updated_at": "2026-07-22T23:07:46.000000Z"
                                        },
                                        {
                                            "id": 5,
                                            "nome": "Dr. Ricardo Nunes",
                                            "crm": "CRM005",
                                            "created_at": "2026-07-22T23:07:46.000000Z",
                                            "updated_at": "2026-07-22T23:07:46.000000Z"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Médicos"
                ],
                "summary": "Cadastra um novo médico",
                "description": "Cria um novo médico vinculando uma ou mais especialidades.",
                "operationId": "9545d280519dd0c08a97b6884b48ebc5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome",
                                    "crm",
                                    "especialidades"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "Dr. João Silva"
                                    },
                                    "crm": {
                                        "type": "string",
                                        "example": "123456"
                                    },
                                    "especialidades": {
                                        "description": "IDs das especialidades separados por vírgula",
                                        "type": "string",
                                        "example": "1,3,5"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Médico cadastrado com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "message": "Médico cadastrado com sucesso.",
                                    "data": {
                                        "id": 1,
                                        "nome": "Dr. João Silva",
                                        "crm": "123456"
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação."
                    }
                }
            }
        },
        "/api/medicos/{id}": {
            "get": {
                "tags": [
                    "Médicos"
                ],
                "summary": "Busca um médico pelo ID",
                "description": "Retorna um médico específico pelo identificador.",
                "operationId": "3489ed5ddbc13d9771aca97ff753b919",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do médico",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Médico encontrado."
                    },
                    "404": {
                        "description": "Médico não encontrado."
                    }
                }
            },
            "put": {
                "tags": [
                    "Médicos"
                ],
                "summary": "Atualiza um médico",
                "description": "Atualiza os dados de um médico existente e suas especialidades.",
                "operationId": "c6c8337575f066f2ebad1ff5159670f5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do médico",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "Dr. Carlos Souza"
                                    },
                                    "crm": {
                                        "type": "string",
                                        "example": "654321"
                                    },
                                    "especialidades": {
                                        "description": "IDs das especialidades separados por vírgula",
                                        "type": "string",
                                        "example": "2,4"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Médico atualizado com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "message": "Médico atualizado com sucesso.",
                                    "data": {
                                        "id": 1,
                                        "nome": "Dr. Carlos Souza",
                                        "crm": "654321"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Médico não encontrado."
                    },
                    "422": {
                        "description": "Erro de validação."
                    }
                }
            },
            "delete": {
                "tags": [
                    "Médicos"
                ],
                "summary": "Remove um médico",
                "description": "Exclui um médico pelo ID.",
                "operationId": "d529b0db449cbfca8443563bb82ff01d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do médico",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Médico removido com sucesso."
                    },
                    "404": {
                        "description": "Médico não encontrado."
                    }
                }
            }
        },
        "/api/medicos/search": {
            "get": {
                "tags": [
                    "Médicos"
                ],
                "summary": "Pesquisa médicos",
                "description": "Busca médicos pelo termo informado.",
                "operationId": "21fd7b1a3fe5a024494b3bc9e03a752c",
                "parameters": [
                    {
                        "name": "query",
                        "in": "query",
                        "description": "Nome ou termo de pesquisa",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "João"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resultado da pesquisa."
                    }
                }
            }
        },
...
    },
...
}